refactor: move history view into src/history.rs
