+++ src/main.rs
@@ -11,6 +11,7 @@ mod divider;
 pub(crate) use divider::{full_island_w, Divider, DIFF_GUTTER_W};
 mod ui;
 pub(crate) use ui::*;
+mod history;
 use kyde_diff as diff;
 mod editor;
 use kyde_config::keymap;
